// src/office/types.ts
export const CharacterState = {
  IDLE: 'idle',
  WALK: 'walk',
  TYPE: 'type',

  SIT: 'sit',
  TALK: 'talk',
  EAT: 'eat',
  DRINK: 'drink',
  COFFEE: 'coffee',
  SHOP: 'shop',
  MEETING: 'meeting',

  PHONE: 'phone',
  READ: 'read',
  THINK: 'think',

  REST: 'rest',
  SLEEP: 'sleep',
  YAWN: 'yawn',

  WUDHU: 'wudhu',
  PRAY: 'pray',

  CARRY: 'carry',
  CLEAN: 'clean',

  ENTER_BUILDING: 'enter_building',
  EXIT_BUILDING: 'exit_building',

  WAITING: 'waiting',
} as const;
export type CharacterState = (typeof CharacterState)[keyof typeof CharacterState];